Start of periodic update / dry-run framework
Spin out the clock and updating aspects of FakeConn to its own module.
This enhances the dry-run functionality by ensuring that there is only
one master clock, which can be either real or fake. In particular,
sleeping has to be centralised to prevent multiple threads from directly
updating the fake clock while sleeping in parallel (i.e. when the script
and update threads both sleep at the same time). Invent a dormitory
structure to achieve this.