Reduce Dormitory to single Bed and merge clocks
We only expect two threads to interact in the fake observation session:
the main thread running the observation script and the periodic updater
thread emulating the KAT system. We therefore do not need a full
Dormitory of sleeping threads but only a single Bed.

The Real and Fake clocks are also merged into a WarpClock. The idea is
that the updater thread will only warp ahead if it is a dry run and
the main thread is sleeping (i.e. is in Bed). If it warps ahead without
regard for the other thread, the main thread will probably get confused
if it finds the system changing underneath it like that during its
non-sleep phases. This will also ensure that the final system time
actually matches the expected script end time during a non-dry run.