Split clock out of updater thread
The updater thread used to instantiate and contain the one and only
master clock to ensure that it is the only thread that can manipulate
time. This created a chicken-and-egg situation in FakeConn, as the
FakeClients need the clock to get built while the updater thread needs
the constructed FakeClients for their component models.

The solution is to split the clock and updater thread functionality.
First, extend the WarpClock to contain the Bed and to handle the
master-slave interaction, thereby making it a standalone time source
that may warp ahead if both the master and slave threads sleep.
Second, instantiate the clock at the top level and pass it to the
updater thread constructor (trusting that no-one else fiddles with it!).