Ensure that we can stop updater thread
Use context management to clean up the updater thread instead of
an atexit / sys.excepthook quitter a la katcorelib as this is more
reliable and well suited to script use (which already has the KAT
object as a context manager). With interactive use the user has to
call kat.stop() explicitly.