Add SingleThreadLock to protect warp clock
A SingleThreadLock only ever allows a single thread to use it. As soon
as another thread attempts to acquire it (or the same thread re-enters
it), an exception is raised. The lock acquisition will therefore never
block. This class ensures that the warp clock will only ever have a
single master and slave thread to avoid clock screw-ups.