Add ClientGroup class and improve Client.wait
A ClientGroup represents a group of clients (i.e. the old Array class
from katcorelib, but with a better name). It passes requests to all
its members via a GroupRequest object. It provides a wait() that
returns once all its members satisfy a sensor condition. The group
is configured in the config file.

In the process, also fix single-client wait() to abort if there is no
strategy set on the sensor. The latest katcorelib has worked around
this case, but revert to older behaviour as it is cleaner. Expose
the strategy name on the sensor object.