Add support for aggregate sensors
Aggregate sensors update their values according to a rule based on
other sensors. This update provides a simple mechanism that mimics
katcp.AggregateSensorTree without adding overhead to the FakeModel
namespace. If a Model has a method named '_aggregate_<sensorname>',
it is assumed that this sensor is an aggregate sensor and this method
is the associated rule, while the arguments of the method are the
sensors providing input values to the rule. The aggregate sensor is
then updated whenever one of the input (or 'child') sensors are
updated (even if its value does not change). The order of events are
important: all the Model attributes have to be initialised first
before the aggregate sensors are registered to avoid a chicken-and-egg
situation.