Fix matplotlib import
The old script had the poor approach of conditionally importing matplotlib
midway through the script. This broke when the script commands got turned into
a function, effectively hiding the import from the rest of the functions.

The fix is to import matplotlib upfront if possible and check that it is
there once it is actually needed.