Allow the use of existing weights if no cal target
Rework the script so that calibrator targets are optional. The script now
picks the closest cal target (if any) that is also up. If no cal target is
provided, use the latest beamformer weights if available (but warn if they
are older than 2 hours). Avoid the use of natural / unity weights as these
will point the beam in some unspecified direction and are therefore as good
as no weights.

Remove the 'phase-once' option as it only meant something with more than one
target, while the script only allows for one target anyway.

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katscripts/trunk@22469 b56c7a7d-ac47-0410-a905-b6f7de78896a