Rework the style in which bf weights are phased up
Consolidate the way gain corrections are turned into beamformer weights.
This is now specified by a single 'style' option which replaces and extends
the 'fix-amp' option. The old default used to be 'norm' while fix-amp toggled
the style to 'full'. The new default is 'flatten', which is like 'full' but
preserves the average gain in each input.

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katscripts/trunk@22470 b56c7a7d-ac47-0410-a905-b6f7de78896a