Normalise beamformer weights by number of inputs
Divide the beamformer weights by the number of inputs to help prevent the
beamformer output from overflowing. This is because both the F-engine output
(and therefore beamformer input) and the beamformer output are treated as
fixed-point numbers between -1 and 1. The sum should therefore not be able
to grow past 1 in amplitude. This might require further finetuning across
inputs to be sure...

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katscripts/trunk@22483 b56c7a7d-ac47-0410-a905-b6f7de78896a