New raster scan --style option
Unify all the options and flags influencing the raster scan style into
a single option called --style. This sets the number of scans, scan duration,
scan extent and scan spacing, and also the default dump rate. The old options
are now deprecated and the user is warned when using them (although they still
work as expected for now).