Revert "Override target on tracking antennas"
This change is considered harmful because in the usual case there is only
one target anyway and the tracking and scanning antennas now slew to the
target one after the other, wasting a lot of time. The reverted commit was
introduced as a quick hack to get holography going in the absence of proper
pointing, which has since been remedied.