Fix baseline indexing bug in baseline reduction
The script made an assumption that katdal data.ants and data.inputs line up.
This is not guaranteed, however (in general they are not even the same size).
Rather base the input list directly on data.ants. This makes the script work
for more than one baseline (and validates undoing the delay tracking).