Add --reconfigure-sdp option
The SDP containers are brought up on product configure and stays up until
product deconfigure. While this allows re-use and quicker setup times, some
containers like katsdpcal do not yet like this setup, expecting to be
resurrected at the start of each observation (i.e. at the capture init/done
stage). This option refreshes the SDP subsystem by deconfiguring and
reconfiguring with the same parameters.

Since katsdpcal is mostly only involved in beamformer phase-up, this seems
like the appropriate place for this functionality for now. It is more
generally useful, though, and could still migrate to CaptureSession or even
the subarray manager. Consider it a hack for now.