Make beamformer weights = 1 / sqrt(num_ants)
This attempts to keep the output power level of the beamformer
independent of the number of antennas being combined, which is useful
for the final 8-bit requantisation step and subsequent processing
to e.g. filterbank files. It assumes that the antenna signals are
mostly uncorrelated (i.e. SEFD >> signal power). Otherwise, the scale
factor should be closer to 1 / num_ants. Also report what we are doing
and check the result of the CBF command (could fail in future).