Avoid numpy 1.12.0 as it fails to build extension
The gsm extension module fails to build with numpy 1.12.0 as there
was a regression in f2py (see numpy issue #8479 and subsequent fix
in PR #8494). It fails to dimension arrays with parameter-based
shapes. The end result is that the package fails to install.

The proposed solution is to add a known good numpy as a setup
(and "pre-setup") require. If your system already has a proper numpy
it is used; otherwise numpy 1.11.3 or a later version will be
downloaded locally for the extension build (and then discarded).