Fix crash and improve reported compscan count
The debug print accesses main_compscan.beam.is_valid even if the beam
object is None, resulting in an exception. Improve that bit of code by
switching to a logger and checking main_compscan.beam first. Also try
to get keep and keep_all to be Booleans as opposed to None (not quite
successful yet).

The dataset in reduce_compscan_with_uncertainty is restricted to a
single compscan early on. The original dataset, however, has not been
selected for scans, which overestimates the number of compscans.
The cleanest option is to pass in the number of compscans to fix the
progress report. It should match the GUI text label.