Disable matplotlib interactive mode
This allows the script to be run from within the standard Python
interpreter (otherwise it would not block at the end of the script
and instead immediately return to the command line).

Coincidentally the occasional button freezes as mentioned in MRTS-410
are now gone, but the spectrogram button has stopped working...
Running from within IPython stills shows the old behaviour.

This addresses Jira ticket MRTS-410 (sort of).