Fix phase-up for AR 1.5
With the latest array release the cbf_center_freq telstate key
(used in the katsdpcal pipeline) changed from the usual 1284 MHz
to 428 MHz (now the centre of the base band as opposed to the actual
band). This meant that a different set of frequencies was used when
correcting the residual delays (i.e. applying the K solutions) in
the pipeline vs in the phase-up script. The end result is large
phase offsets on each input due to the different fringe phases
in the two scenarios. Kudos to Laura for spotting this!

The fix uses the cal channel frequencies from telstate instead of
synthesising them ourselves (which was very hacky to start with).
It would be even better to use the latest delay adjustment framework
in the CBF proxy but this requires the cal pipeline to get proper
sky frequencies first. Refuse to fix delays without the cal version
of these frequencies.

Also clean up some cruft.

This addresses JIRA ticket MKAIV-543.