New delay_cal observation script
This is an offshoot of the original bf_phaseup_AR1 script that only
focuses on delay calibration. Unlike the phase-up script the delay
adjustments are done to the delay tracker in the CBF proxy as opposed
to synthesising a phase slope to apply to the F-engine gains. It can
therefore properly account for coarse delay corrections (longer than
one sample) instead of being restricted to fine delay corrections.