Set F-engine gains if requested
The calibrate_delays script is intended to be the first script run
on a new subarray and it is therefore convenient to set the default
F-engine gain here too. In future this feature might move to the
subarray controller but this is a good place to start.

The gain can either be specified directly via the --fengine-gain
option, automatically determined from the channel count (the default)
or its setting can be skipped by specifying a negative value (useful
if this script succeeds e.g. phase-up for some reason).

This functionality used to be seated in bf_phaseup_AR1 but that script
is falling out of favour for initialisation as it inadvertently
imprints properties of the calibrator source onto the system gains
which is not great for imaging.