Use bfcal tag for delay calibration (for now)
The seemingly more appropriate 'delaycal' tag triggers a preliminary
gain calibration followed by an actual delay calibration in the
standard cal pipeline. This preliminary calibration can fail on some
antennas, resulting in large residual delays.

Disable it implicitly by reverting to the 'bfcal' tag. This is
overkill, as the pipeline then also produces bandpass and gain
solutions, but at least ensures that the first step is pure delay
calibration (and since solutions are quickly available the timing is
still the same).

A better solution would be to fix the delaycal tag or make yet another
custom tag.

This closes JIRA ticket MKAIV-594.