Extend horizon check to entire compound scan
Attempt to check horizon in a similar fashion to session.raster_scan().
First lift the horizon so that the lowest pointing (which is max_extent
below the target in elevation) will also be up if the target is up.
Then check if the target is up for the entire duration of the compound
scan (one track on target + pointings in x + pointings in y).

As suggested by Lance, also remove explicit set_target() calls as
session.track() already takes care of this.