Discard first and last gain chunks
These chunks on the band edges show higher variances, not because of
intrinsic variability but rather due to large systematic changes as
the gains drop off. Even though their weights will be low, discard
them anyway as we have enough in the middle.

Also do a slight refactor of old BeamPatternFit code to reduce both
number of lines, line length and repetition.