Reduce sleeps and track durations
The flashy new pipeline only takes a few seconds max to process the
full bfcal. Coupled with some data lugging overhead (on the order of
17 seconds), this suggests that 60 seconds is more than enough to
wait for gains to materialise. It is also possible to track for
shorter durations (minimum 2 whole dumps), so adjust the default
script options accordingly (and keep it multiples of 8-second dumps).