Move horizon check into session block (and dedent)
This ensures that the requested sources are checked to be above the
horizon during the dry run / schedule block verification. Previously
the check was done before session instantiation and therefore used
the unadulterated time.time(). By moving it inside the with-session
block the check uses the appropriate time.time() and will therefore
be valid even during dry runs.

These older scripts also now gain the NoTargetsUpError exception,
which allows a lot of code to be dedented by one level.

This addresses JIRA ticket CAM-513.