Remove noise diode splits and fix horizon check
There is a bug in the horizon checking where the session attempts to
track a target that is busy setting, but then ignores the failed track
and attempts this ad infinitum. In a real observation the target
eventually sets and the script quits. A dry-run, however, gets stuck
on the same timestamp and generates reams of logs.

The solution involves a general cleanup of the tracking logic.
The automatic splitting of tracks into noise diode periods has been
removed, as this obfuscates a lot of the logic. By selecting a track
duration commensurate with the noise diode period, most of the script
functionality is preserved. The only issue is with multiple targets
having multiple noise diode firings per track each.

This addresses JIRA ticket CB-2222.