Replace explicit time.sleep with telstate.wait_key
This allows the script to continue as soon as the cal pipeline products
materialise in telstate, instead of sleeping for the worst case time.
Consolidate all product retrievals in a single routine (DRY).

Use the same delaycal routine for both phaseup and calibrate_delays,
and resign ourselves to the fact that katsdpcal will forever have delay
solutions with the wrong sign (which at least makes interpretation of
historic solutions sane).

This addresses JIRA ticket SR-965.