Correct HV delays during delay calibration
This commit fires the noise diode during the observation of the delay calibrator. If the noise diode is fired during a 'bfcal' tagged track the pipeline will solve for 'KCROSS_DIODE' and put this into telstate.

It is NB for the pipeline that the noise diode is fired before the track starts and remains on for the entire duration of the track.

The HV delays are added to the original H and V delays relative to the reference antenna before being applied.