New --verify-duration option
Make the time spent revisiting the source to verify phase-up
configurable. The default is 64 seconds which ensures a few dumps
for eyeballing the signal displays but not too much. A duration of 0
or less disables the revisit to save time. This was previously stuck
at the same value as --track-duration, which could be up to 10 minutes
for accurate bpcals.

This restores and extends the intention of PR #278.