Warn about NaNs in delays and then ignore them
The delay fitter in the cal pipeline returns a NaN if a delay solution
could not be found, typically because all the data of the relevant
correlator input was flagged for some reason. Discard those solutions,
leaving the delay adjustment at zero for the affected inputs.

Crash if *all* delays are NaNs, though. That's just not on.