Use catalogue from katexternal data of gps scans
The `gps-opts.txt` catalogue file now lives in katexternaldata on the
monctl node and is updated automatically.  Changed observation scripts
to get the file contents via the katconf server instead, as the file is
no longer available on the obs node.

Requested by Sulayman Salie for track_scan_drift.py.