Update phase up default requant gains
This is an update to default requant gains dict with approriate requant gain values that have been determined from mesurements. The 1k has a value of 116 for a FFT shift of 255 the 4k has a gain value of 70 for a FFT shift of 511 anf the 32K has a gain of 360 for a FFT of 511.