Deprecate --default-gain, use --fengine-gain
The calibrate_delays and bf_phaseup scripts have drifted out of sync
a bit, e.g. by using different command-line options for the same thing.

Use the more modern name for the option setting the gain. The 'default'
gain is the automatically determined gain based on the number of
channels in the F-engine, while this option intends to override that.
Also calling it --default-gain is therefore confusing.

Support both names in bf_phaseup and let --default-gain still override
the new option, but warn the user that the old option is deprecated and
may therefore disappear in future.

Also fix the help string of the --fengine-gain option to reflect the
truth (-1 is actually treated the same as 0, not a special case that
leaves the gains alone anymore).