Simplify bandpass cleanup to go with cal pipeline
The recent PR ska-sa/katsdppipelines#330 already cleans up the bandpass
solutions in the cal pipeline, also extending flags to adjacent channels
and applying them to both H and V like this PR used to do. In addition,
proper flagging of the BP solutions means that the crude median filter
should not be necessary anymore.

The remaining clean-up steps in the script are now linear interpolation
and flagging of large gaps and band edges.