Improve gain magnitude normalisation
Some history first...

The original phase-up script had a fixed magnitude for the F-engine
gains (aka gain corrections) that was constant for all inputs and
frequency channels (let's say 70 for the purpose of this commit
message). It really only adjusted the phases per input and per channel
to phase up the beamformer:

  (1)  corrections = 70 * abs(gains) / gains

The first version of --flatten-bandpass (#252, #253, Oct 2017) had

  (2)  corrections = 70 / gains

which flattened the bandpass magnitude but unfortunately also introduced
an arbitrary scaling, since the gains have the arbitrary units of
sqrt(Jy / count). This caused clipping and/or quantisation loss on some
F-engine outputs.

The next version of --flatten-bandpass (#331 redone as #335, Nov 2018)
addressed this by normalising by the mean gain over the band per input:

  (3)  corrections = 70 * average_over_frequency(gains) / gains

While this removed all the bumps in the bandpass of each input, it left
the average correction magnitude stuck at 70 for all inputs, thereby not
equalising the average gain per input. This leads to some H / V mismatch
as well as potential correlator / beamformer inefficiency.

The proposed change addresses this equalisation issue. The gain
correction calculation is collected in its own function to clean up
the main script flow and allow simpler testing. The average gain is now
calculated over all inputs and all frequency channels, using a more
robust succession of medians that ignore flagged gains, as

  (4)  corrections = 70 * global_average(gains) / gains

The average (relative) gain per input is also reported as a diagnostic,
and the script refuses to continue if all the corrections will be zero
(which implies zero signal out of the beamformer).