Safeguard dry run and sort relative gains
The dry run crashes because the `get_cal_solutions` calls return empty
dicts. It would be much more convenient if they returned sensible dummy
values... Until then, stub out the calculation and setting of
corrections (at the cost of changing the script output in dry run,
which is undesirable).

Also sort the relative gains by input to make it easier to spot the
gains of a specific dish. And clarify what a "relative gain" is.