Preserve band edges in beamformer output
The current incarnation of the cal pipeline includes the first 5% and
the last 5% of the band in the static RFI mask, so those edge channels
are always flagged in the BP solutions. The bf_phaseup script then also
did not interpolate over the band edges, which zeroed them out in both
the correlator and beamformer data.

Relax this behaviour by restoring linear interpolation at the band
edges (or constant extrapolation, more likely). This will still tend to
suppress the band edges by adopting a constant gain instead of the
rising gain needed to counter the drop in SNR in these regions. This
should therefore be benign while providing the opportunity to extract
useful info from these channels.