Suppress inputs with out-of-range relative gains
If a particular input has a problem with low SNR, the corresponding
relative gain will be low. This will result in excessive correction
in the F-engine (since correction = 1 / gain) which will boost the
noise and potentially saturate the input signal. The opposite is also
possible for inputs with excessive SNR. Fix this by clipping the
relative gain to the range [0.5, 2.0], since typical relative gains
are in the range [0.8, 1.4].