Handle absent antennas in solutions of ref-point
Receptors might appear in session.observers but be flagged or marked
absent and then not show up in the cal pipeline solutions. Be more
robust if this happens. Also correct the docstring's references to
the local :class:`BeamPatternFit` (no scape in sight!).

This addresses the original concern in JIRA ticket SR-1807.