added "no-corrections" switch to force static and tilt corrections OFF
it seems that static and tilt corrections are applied in "RATE" mode unless explicitly disabled ([OM-10120333-30000-01-01 Rev 1.0], section 3.3.4.5, confirmed by experiment).

since it is often desirable to use this script with all corrections disabled, i have now added such a switch. the default operation of this script remains unchanged. the original state is restored when the script ends successfully. [TODO: original state is hard-coded, need to finalize that part of it).