Pick strongest calibrator instead of highest
This will increase the signal-to-noise ratio and lead to better gain
solutions. The flux density has to be evaluated at some frequency, so
I pick the delay-tracking centre frequency returned by the session.
This still has to be checked for suitability, especially for UHF
(the dry run has a hardcoded frequency of 1284 MHz, for example...).