Make calibration selection more robust
If the flux density could not be determined for any source (typically
because none of them have flux density models or the flux frequency is
way out), rather fall back to the source with highest elevation with
a warning. This makes it easy to phase up on new sources without
mandating a flux model first.