Fix catalogue selection
First filter the sources to get all those that are above the horizon,
and then extract the ones with valid fluxes. Otherwise, a strong source
below the horizon will win as long as at least one source is up (oops!).
Move some lines of code to more appropriate spots closer to where they
are applicable.