Pick the first calibrator that is up
Tweak the calibrator selection logic in the phase-up scripts.
The previous scheme picked the source with highest flux that is up.
The problem is that J1939-6342 is preferred to J0408-6545, but it is
weaker in the middle of the band. Therefore rather use the existing
ordering in the catalogue to indicate priority (the catalogue filters
also preserve the order). The catalogue should be ordered from the
source with highest priority (i.e. J1939-6342) to the source with
lowest priority (i.e. J0825-5010).