Upgrade F-engine interaction in calibrate_delays
Remove default F-engine gains stored in the script and rely on the
new 'default' gain feature in CBF instead, as supported by the recent
ska-sa/katcorelib#325 PR. The same goes for the FFT shift.

Gain and shift are set to 'default' by the --fengine-gain and
--fft-shift script options, respectively, but can still be overridden.
These options are of the new 'int_or_default' type introduced by the
katcorelib PR.

Use the new `adjust_fengine_delays` method as well. This takes care of
any NaNs and also provides a clean and simple way to zero the delays.
Improve docs and comments.

This addresses JIRA ticket SR-1952.