First reset F-engine to a known good state
Before checking for the presence of the cal pipeline and a visible
target, first set the F-engine gains / FFT shift / delay adjustments.
This way, the F-engine can be set up even without cal or a target.
These tests will eventually move to CaptureSession, but now they are
at least in the correct spot in the script. Also normalise the two
scripts a bit by moving `collect_targets` to the same spot.