Apply BCROSS_DIODE aka HV phase corrections
Fire the noise diodes as part of phase-up (both initially and during
verification). This triggers the cal pipeline to produce KCROSS_DIODE
and BCROSS_DIODE solutions. Get those solutions from telstate and
combine them with the K and B solutions, respectively. Also do a few
tweaks to the script logger output to synchronise this script with
calibrate_delays, and remove somewhat spurious logs.

This addresses JIRA tickets SR-1940 and SR-1980.