Start all available correlators instead of The One
This supports multiple instruments aka narrowband. We are not starting
the beamformer streams but at least all the correlator streams.
The bare minimum would be to start the wide instrument only, since that
is all that is needed by cal. This way we get some regular narrowband
data on calibrators too, which could be very useful in the archive.

This addresses JIRA ticket SR-1841.