Start all CBF streams in observation scripts
After some discussion with the beamformer folks it turns out to be OK
to start the beamformer during calibration observations, since the
PTUSE et al will only start capturing after a ?target-start request,
and those are only called in the beamformer scripts.

This simplifies the logic in the scripts and avoids a somewhat awkward
parameter to `session.capture_start()` to specify which streams to
start (correlators only? what about FBFUSE? etc etc). There is also no
need for a separate `session.capture_init()` since that is also done by
capture_start.