Trim target tags up front to eliminate "multiple targets"
Current practice is for target tags to be trimmed to "avoid overloading the cal pipeline". Until now this trimming was done after the target was used once (without trimming). This results in the target appearing twice in the catalogue (un-trimmed & trimmed).

I have simply moved the trimming to be done before the target is used the first time.