Apply BCROSS_DIODE_SKY solution
This is the full HV phase correction that includes the transform of
the noise diode HV phase to that of the sky (derived on 3C286). This
transform is implemented in the cal pipeline as a band-specific spline
that is common to all antennas. The BCROSS_DIODE_SKY solution combines
the median BCROSS_DIODE solution with this spline.

If BCROSS_DIODE_SKY is unavailable (most likely because no-one has
derived a noise-diode-to-sky transform for the current band yet),
abandon all HV phase corrections. This ensures that we only have two
scenarios in the archive: no HV corrections during phase-up, or full
HV correction. Applying BCROSS_DIODE only would introduce a third
scenario of halfway correction.

This addresses JIRA ticket SR-1980.