Bug fix for targets with & without tags
This code has been defective if a compound scan has multiple targets, even if they differ just based on the number of tags. This is the case with the current and historical versions of observation/interferometric_pointing.py but this only becomes apparent when the number of measured points are reduced. See also PR #417 .

The change is to test for equivalence of targets EXCLUDING tags, and then to select the scans for all of those targets. Previously the code would typically have omitted the first scan, which is the one where the target has more tags than the rest of the scans in the compound scan.