Add --disable-hv-correction and fallback
The new --disable-hv-correction script option still fires the noise
diode to get the full cal solutions into telstate, but then don't apply
the HV bandpass and delay corrections. This is only really useful for
testing the script against other HV correction methods, as requested
by Maciej.

If BCROSS_DIODE_SKY is not available, fall back to BCROSS_DIODE. This
therefore fixes the electronics part of the HV phase, but not the
fixed OMT part relating HV phase to the sky. The typical reason for a
lack of BCROSS_DIODE_SKY is the use of a new band like UHF and S, which
does not yet have an associated OMT spline. Previously we disabled
HV correction in this case, to make it easier to piece together which
gains were applied to the data. Since the new bands are rarely used
with phase-up yet, it is not as onerous as previously thought, and it
provides the majority of HV correction in that case. We could also
prime new bands with receiver-provided curves for starters, making
a lack of BCROSS_DIODE_SKY less likely.