Check that target is up for duration of raster
Added straight-forward code to check that the source remains above horizon for the estimated duration of the raster.

To facilitate this, I had to re-factor the code to separate (A) selection of the raster scan pattern from (B) the execution of the selected pattern. This change does not affect any of the scan patterns or other logic.