SPR1-407 Fix figure title to handle many datasets
Previously the list of datasets clobbered the rest of the labels when
the list grew too large. Cap the list at 4 entries and always display
the first and last entry, since those will typically indicate the date
range of the data going into the fit. Space out the version string,
antenna name and dataset names in a single text label, thereby
preventing any overlap in future.