SPR1-407 Allow fixed parameters during fitting
Allow pointing model parameters to be fixed to their old values during
fitting, based on ska-sa/katpoint#73. Introduce three parameter states:

 - ZEROED: the parameter is zero and won't be fitted
 - ENABLED: the parameter will be fitted to the data
 - FIXED: the parameter wil keep its old value and won't be fitted

The buttons cycle through these states, where FIXED is indicated by a
dark frame around the button and ENABLED is indicated by a bold
parameter name. The states are implemented with plain ints instead of
enums because they are easier to cycle (and Python 2.7 still needs an
extra enum34 package for enums).

Use the opportunity to get rid of some conditional expressions and to
consolidate the creation of parameter strings.