SPR1-1418 New beamformer raster scan script
This electronically scans the beamformer across a source in a series
of pointings. Each pointing lasts at least long as the delay model
update rate, to ensure that it is properly set. This is typically used
in conjunction with a custom SPEAD receiver that dumps the beamformer
ingest signal display stream (1 Hz dump rate) to a file.

It currently only steers beam `0x`, since the signal display only
supports `0x` and `0y` anyway. This is the latest version used in the
steerable beam tests on 22 September 2021.