added active_ants() to eliminate dead antennas by default
if any of the "selected" ants is stuck, i.e. doesn't track with the others, then NO solutions are generated for any of the scans.

previous situation was that the user has to somehow determine which antennas to exclude.


i've also added more doc comments for 'activity()' and deleted a redundant log message.
with this change the new 'active_ants()' is used to set the default list of ants to automatically exclude antennas that never track.