SPR1-2393 New geodetic block scan script
This tracks many (e.g. 10 to 30) well-spaced targets in a short time
(e.g. 30 minutes), which is useful for delay/baseline calibration
because it ensures reasonably constant atmospheric effects, which may
be fitted and removed from the delays.

This is known as a "geodetic block" observation in VLBI, hence the name.

It also features a funky new "bitonic" scheduler that tries to scan
as many targets as efficiently as possible in the given time.