SPR1-2393 Stay 10 degrees away from Sun and Moon
This avoids occultations and other nasty business. The distance is
based on solar interference estimates up to 7.5 degrees away from
boresight with the S-band feeds, plus a bit of leeway to make up for
the movement of the non-celestial sources. The distance is measured
in the middle of the scan.