Upgrade of refine attenuation to have beeter reports and work with UHF
The Refine attenuation code measures the standard deviation of the ADC samples is measured for both the UHF and L-band and a threshold of 12 is used. This is done to ensure there is enough timing information to maintain the receiver sampling rate. The voltage and voltage range thresholds are are measured from the ADC samples and a suitable range is 190 and 190+ 20 for L-band, and √2 x 𝐿𝑏𝑎𝑛𝑑 and x for the UHF-band. Given (𝑣𝑜𝑙𝑡𝑎𝑔𝑒) √2 𝐿𝑏𝑎𝑛𝑑(𝑣𝑜𝑙𝑡𝑎𝑔𝑒 𝑟𝑎𝑛𝑔𝑒) the threshold, if the antenna’s standard deviation and voltage are not within the acceptable range, the algorithm changes or updates the attenuation levels until the values lie within the appropriate range. The end of the refine attenuation process is two results emails sent to the operators. The first email is titled “Summary: Changing attenuation”, this email summarizes the entire observation and is important to be seen by the operators as it highlights the antennas that have changed, and should be updated in the config. The second email is titled “Details: Changing attenuation”, this email contains all the debug information, it highlights all the detailed changes or updates made during the observation and is useful for a deeper understanding of what happened during the observation.