Simplify Pointing Code and change projection
The rework of the code is to make it simpler with changes to split the main routine reduce_compscan into two parts the fit beam part and all the assoated field parts .  The second part has been to change the projection to use swapped sin projection .  Since this should work better since the beam is scan around the target . This is done twice in the code . the first time with getting the x & y offsets from katdal and the second from the calculation of the beam offset. this cancels out the swapped perspective.  but the refraction is calculated from the pespective of the scanning beam not the target. Fix the np.bool warning and tidy code