Implement an antenna/timstamp flagging.
This change involves moving the antenna activity select into the compscan routine. Larger changes were to change the activity check to rather return not the times when all antennas were tracking but an array of antenna,timestamp of the antenna tracking. The rest of the changes are to account for this in all the means and weights that are used. This was needed as antennas sometimes fail to lock and this would casue that whole scan to be marked as, not valid. Since we only have 5 tracks with 4 values fitted , the fits fail if one track is lost due to this.  this will hopefully enable better fitting of bad areas of sky such as low elevation areas .