Store refracted azel in CSV output
The analyse_interferometric_pointing script currently stores the
"actual scan" (az, el) coordinates as "azimuth" and "elevation" in
the output CSV file. This is pre-refraction, i.e. topocentric / out in
space. However, these values are supposed to be what's fed into the
pointing model in order to fit the offsets appropriately when passed
to the fit_pointing_model script, so it should include refraction.

Therefore change the stored values to the "refracted scan" coordinates,
i.e. the observed / post-refraction / on-the-ground (az, el) values.
This matches the behaviour in the analyse_point_source_scans and
reference_pointing scripts.

This used to be correct but a change sneaked in as part of the very
first PR 8 years ago, and I missed it :-)